2023년 4월 8일 토요일까지 작업 내용 ( 토스 페이먼츠 테스트 결제 API 연동 )

+++ src/router/user.ts
@@ -6,46 +6,32 @@ import {
   readDataBase,
   readDataBaseJoin,
 } from "../database/database";
-import { getUser } from "../controller/userController";
+import axios from "axios";
+// import { getUser } from "../controller/userController";
 
 const userRouter = express.Router();
 
-userRouter.get("/", getUser);
+// userRouter.get("/", getUser);
 
 userRouter.post("/", async (req: Request, res: Response) => {
   const to_db = {
     email: req.body.email,
     name: req.body.name,
-    position: req.body.position,
-    user_type: req.body.position,
-    play_type: req.body.position,
     point: 0,
     level: 0,
     password: bycript.hashSync(req.body.password, 8),
   };
   const email_auth: any = await readDataBase(
     "email_auth",
-    ["email", "auth_num"],
-    `email='${to_db.email}'`
+    ["is_auth"],
+    `email='${req.body.email}'`
   );
-  const fullReselt = email_auth[0][0];
-  if (fullReselt.length === 0) return;
-  if (fullReselt.auth_num === req.body.auth_num) {
-    try {
-      await createDataBase("user", to_db);
-      await deleteDataBase("email_auth", `email='${to_db.email}'`);
-      return res.status(200).json({
-        result: true,
-        message: "회원 가입을 완료했습니다.",
-      });
-    } catch (error) {
-      console.log(error);
-    }
-  } else
-    return res.status(400).json({
+  if (!email_auth[0][0].is_auth) {
+    return res.status(200).json({
       result: true,
-      message: "인증번호가 잘못되었습니다.",
+      message: "이메일 인증을 완료해주세요.",
     });
+  }
   try {
     await createDataBase("user", to_db);
     await deleteDataBase("email_auth", `email='${to_db.email}'`);
@@ -54,7 +40,10 @@ userRouter.post("/", async (req: Request, res: Response) => {
       message: "회원 가입을 완료했습니다.",
     });
   } catch (error) {
-    console.log(error);
+    return res.status(200).json({
+      result: true,
+      message: "회원 가입에 실패했습니다.",
+    });
   }
 });
